Character クラスを分割、ファイルの分割
