+++ main.rb
@@ -1,17 +1,5 @@
 require 'pry'
-
-# 親クラス
-class Character
-  attr_reader :offense, :defense
-  attr_accessor :hp, :name
-
-  def initialize(**params)
-    @name = params[:name]
-    @hp = params[:hp]
-    @offense = params[:offense]
-    @defense = params[:defense]
-  end
-end
+require './character'
 
 class Brave < Character
   SPECIAL_ATTACK_CONSTANT = 1.5
